fix: Nettoyage du context
